audio-generator/pyproject.toml

[project]
name = "audio-generator"
version = "0.1.0"
description = ""
authors = [
    {name = "camille-vanhoffelen", email = "camille-vanhoffelen@users.noreply.github.com"},
]
dependencies = [
    "suno-bark @ git+https://github.com/suno-ai/bark.git",
    "nltk>=3.8.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}
[project.optional-dependencies]
ide = [
    "setuptools>=67.8.0",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
